Resolve conflicts against master's convergent fixes:

- ShardedModelBase: take master's InvalidateLayoutState(), a virtual
  superset of this branch's `CachedFullParameters = null` that DDP, FSDP,
  Hybrid and PipelineParallel each override to clear their own layout.
- ShardedOptimizerBase, LayerCloning: comment-only conflicts; keep
  master's wording so this branch does not reword master's fixes.
- Bucket12_DistributedTests: keep this branch's stricter contract (a
  downstream exception is no longer a pass, plus prediction sanity
  checks) on top of master's fixture isolation -- the backend now takes
  DistributedEnvironmentId and is registered via OwnCommunicationBackend,
  which owns Shutdown at fixture teardown and so replaces the branch's
  `finally { backend.Shutdown(); }`.

+++ src/NeuralNetworks/AudioVisualEventLocalizationNetwork.cs
@@ -1323,7 +1323,7 @@ public partial class AudioVisualEventLocalizationNetwork<T> : MultimodalModelLay
             }
         }
 
-        return _numOps.Divide(totalDiff, _numOps.FromDouble(frames.Count * frames[0].ToVector().Length));
+        return _numOps.Divide(totalDiff, _numOps.FromDouble((double)frames.Count * frames[0].ToVector().Length));
     }
 
     private Vector<T> EncodeTextDescription(string description)
